deb: dbconfig: don't set dbserver, if it is localhost
When communicating with a local PostgreSQL database server,
it is advised to use a socket connection instead of TCP.
When dbaddress is not set in the Bareos Catalog configuration,
Bareos tries connects via sockets. If it is set, it connects via TCP.
As dbconfig-common by default specifies "localhost"
for a local PostgreSQL database, we ignore this setting.