RPM: extra restart on upgrade from < 25
On upgrade, the bareos-fd will be restarted (in %postun of the old
package).
When upgrading from Bareos < 25, we need a workaround to keep the
existing configuration files.
However, during the upgrade, some files get renamed
(because we can't prevent RPM from doing so).
We restore these files in %posttrans.
As another workaround,
we check if bareos-fd is running %pre.
If yes, we will start it again in %postrans.

This is only required for bareos-fd, as the other services don't got
restarted.