cmake: fix traymonitor
When we changed how we use vcpkg (-> manifest mode) we did not update
the windows installer to use the new paths.

To prevent problems in the future we now take everything (except for
the openssl executable) directly from the cmake binary directory.

This is possible because qtdeploy moves everything needed by qt to
that directory.  In the future we should look into qt cmake support
which would do this in addition to other things automatically.

(cherry picked from commit 1fca2de7612421e8cf529cf0837deddc02b00918)