debugedit: fix messing with mtime
Changing the mtime of a library _after_ the .so symlinks were created
messes with cmakes internal logic. In those cases cmake will decide to
relink the library itself, meaning that everytime you execute a build,
all libraries (and all targets depending on them) will get relinked.

This is not a big issue on debug builds, but it takes a long time when
building release builds with lto.

This commit fixes this by using a script that restores the mtime after
debugedit messed with it.