dir: refactor status subscription queries
these used two views in the database previously. As it turned out,
changing these requires a schema upgrade. As this is largely unfeasible,
we use a WITH-clause in the query that declares the same relations the
views have.
This removes the need for the views in the database and thus allows to
change the underlying queries without a schema change.

(cherry picked from commit f5e65f3943267c79581248b319d18e3818e6b311)