VMware Plugin: Fix errors if DC not in root folder
In most cases, the DC (datacenter) object is located in the root folder,
but it could also be inside a folder. In that case the plugin failed to
backup the NVRAM files and on restore it didn't find the DC. Both cases
are fixed now.

(cherry picked from commit f9d5dbf69f82bf2987af45d151b45bc60690d083)