rpm: don't fail, if install log file isn't writable
As our RPM installation handles some complex update scenarios
it write to an install log file.
On some systems, it is not possible to write log files,
because /var is read-only. Specifically this is the case on immutable
systems like Fedora Silverblue. In that case, we write to stdout.

Fix #2465